+++ vp/src/platform/hifive/can.h
@@ -96,8 +96,8 @@ class CAN {
 
 	uint8_t readRxBuf(uint8_t no, uint8_t byte);
 
-	void mcp2515_id_to_buf(const unsigned long id, uint8_t* idField, const bool extended = false);
-	void mcp2515_buf_to_id(unsigned& id, bool& extended, uint8_t* idField);
+	void mcp2515_id_to_buf(const uint32_t id, uint8_t* idField, const bool extended = false);
+	void mcp2515_buf_to_id(uint32_t& id, bool& extended, uint8_t* idField);
 
 	void enqueueIncomingCanFrame(const struct can_frame& frame);
 	void listen();
